Created new Application class, moved episode update into an AsyncTask and added CREDITS.
Created an new PonyExpressApp class extended from Application.  Moved the DbHelper instance into the app class.
Moved the Episode update code into an AsyncTask, so it does not block the UI thread.  A new progress dialog informs the user that an update is happening.
Network connectivity is now checked prior to attempting an update. If no connection is available a Toast informs the user.
Added a CREDITS file.