Created Downloader IntentService and InternetHelper classes.
New Downloader IntentService responsible for downloading and saving podcasts to the sd card.
Downloader is activated by a long click context menu on an episode title.
New InternetHelper class has methods to check the connectivity status of the phone.  This class is called from the PonyExpressApp class to make it available from all activites/services.