The seek bar now handles changes in the activity life cycle correctly.
When the activity is destroyed by a change in orientation then the state of play is saved by onSaveInstanceState.
When the activity is paused and resumed the state is saved and restored.
When the activity is destroyed and not immediatly recreated, the state of the podcast player is queried on binding and if the same title is already playing the seek bar is updated to reflect this.