Refactored PodcastPlayer and PlayerActivity in preparation of displaying elapsed time.
PodcastPlayer now has two media players that are cycled.  One is intiialised when PlayerActivity binds and then it becomes the active player when playback begins.  The other player that may have been playing previously then becomes the free player ready to be init by a new PlayerActivity.
This means that we can now query the non-playing player for info about the episode such as its duration and current position.