Added support for other Sixgun productions podcasts (Rant-o'Fab).
Added a new Podcasts table to the database.  This holds the name of the podcast, the feed url, the albumn art url and the name of the table in which to store the episodes.
Each podcasts episodes are stored in a unique table and the ogg files are now stored in a different folder for each podcast.
EpisodeActivity is now generic and works for any podcast.
All of the methods that act on Episode tables have been modified to also take the name of the podcast as a parameter, in order to act on teh correct table.