UI improvments and a bugfix for a regression introduced in commit 32a9845f.
Album art is now downloaded and displayed with the podcast listing and on the episodes activity.
The title of the Episodes Activity is now the title of the podcast whose episodes are being listed.
Added the sixgun productions logo to the podcasts listing.
BugFix: the database consistency check could not find any downloaded episodes and so was marking then not downloaded in the database.