Settings and Setup Identica account in the menu now work.
Added a PreferencesActivity that is launched from menu->Settings.
Can now set the number of episodes to store and the feed update frequency from there.
IdenticaAccountSetupActivity is now launched from the menu->Setup Identi.ca Account.
This necesitated the move of setCredentials and verifyCredentials calls from IdenticaActivity to IdenticaAccountSetupActivity and making IdenticaAccountSetupActivity able to bind IdenticaHandler.