Initial changes to merge Downloader and Player activities.
Moved Download button to Player activity and deleted DownloadActivity.
Moved Player controls into a seperate RelativeLayout so they can be made to SHOW or GONE together.
If the intent that starts PlayerActivity contains a URL in Extras then the Download button is shown, if not then the player controls are shown.
Fixed bug when showing PlayerActivity in landscape, the controls and seek bar were not visible, so removed album_art when in landscape.