Completed the merge of the Downloader into PlayerActivity. Closes LP:608123.
The download asynctask now reports its progress with a progress bar in the Activity as well as in the notification area.
The progress bar and disabled Download button are maintained when the orientation changes.
When the download completes, it is no longer necessary to leave the activity and return for the UI to become the player, it now happens automatically.