BugFix: Media buttons on headset cable should now control the player on all android versions. Closes LP:611231.
Used reflection to enable the use of the new Android 2.2 API for registering media buttons.
This should make them work properly on devices running 2.2 or greater.  Older devices should still work but may be buggy as the old API was rubbish.
RemoteControlReceiver moved to its onw class and registered in the manifest as required by the new api.
Result of all this is that Pony is now built against Android 2.2 (and therefore targetting 2.2) but should be backward compatible to 1.5.