BugFix: DownloaderService is now started by PlayerActivity as well as bound, to make sure it is not killed erroneously. Closes LP:710563.
On some devices, pressing Home while downloading would kill the DownloaderService and block the UI.  The DownloaderService is now started by the PlayerActivity so should remain running until it kills itself when download is complete.
Also, fixed a seperate bug where the download progress bar was not started when returning to the app, after pressing home.