BugFix: Episodes can now be re-downloaded. CLoses LP:694158.
A long click on an episode brings up a conext menu with re_download as an option if that episode has been downloaded.

Also tidied up the context menu: Added a title listing the episode the menu is for, Added a Listen menu item, and added some logic to dictate which menu items are shown.