BugFix: Fixed a FC on startup and a FC when encountering a corrupt sound file. Closes Bugzilla 23 and LP:785208.
When Pony found files in the PODCAST_PATH that were not in the database it would crash when trying to find them (Bugzilla 23).
Pony would crash when encountering a corrupt sound file. Now the exception is handled and a notification is raised suggesting the user re-download the episode via long pressing the episode. (LP:785208)