BugFix: Feed urls are now checked for a response froma server before being accepted by Pony. Closes BZ:16.
Added urls are now checked to ensure a server responds with a code 200, before they are accepted into the database.
Additionally, the PodcastFeedParser now returns null on failure rather than an empty Podcast.  This can then be checked before use.

Unparseable dates (as in the Tuxradar feed) are now handled better by the insertion of a dummy date (the Epoch).  This stops Pony FC when trying to access missinng dates.
Also, moved the Http resonse code check from BaseFeedParser to Utils.