BugFix: Mp3 files wouldn't resume from the correct location with android 2.2 on some HTC devices. Closes BZ:29.
A bug in the HTC rom for android 2.2 meant that mp3 files would not seek correctly. The player claimed they were seeked but play would resume from the start.
Repeat calls to seekTo the correct location are ignored...
This works around this by seeking to a location a little from where we want and then seeking again to the correct location after playback has started.