Completed DownloadOverviewActivity. Shows progress of all active downloads and provides a cancel button for each.
Also: Added an error notification in the case of the SD card not being writable.
Fixed a FC when the DownloaderService is started with a null intent.
Move the percentage downloaded calculation from DownloaderService to a method in DownloadingEpisode.
Minimum version is now Android 1.6 as Pony is currently broken on 1.5.